+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com).
 ## 0.3.0
 
 ### Added
+- Negative values can be plotted using both `draw_logo` and `draw_line` (as well as `facet_plot`). A new example Jupyter notebook was added for such plotting.
+
+### Added
 - Example notebooks now run on `mybinder`.
 
 ### Changed